requirements.txt
# Node.js backend, not Python. No requirements needed here.
google-generativeai==0.3.0
python-dotenv==1.0.0